refactor(types): colocate listen options with instance
