+++ types/instance.d.ts
@@ -19,7 +19,6 @@ import {
 import { preHandlerAsyncHookHandler, preHandlerHookHandler, preValidationAsyncHookHandler, preValidationHookHandler } from './hooks'
 import { FastifyBaseLogger, FastifyChildLoggerFactory } from './logger'
 import { FastifyInstanceHooks } from './instance-hooks'
-import { FastifyListenOptions } from './instance-lifecycle'
 import { FindMyWayFindResult, PrintRoutesOptions } from './instance-routing'
 import { FastifyRegister } from './register'
 import { FastifyReply, FastifyReplyForRoute } from './reply'
@@ -49,7 +48,28 @@ import {
 
 export type { FindMyWayVersion } from './router-options'
 export type { FindMyWayFindResult, PrintRoutesOptions } from './instance-routing'
-export type { FastifyListenOptions } from './instance-lifecycle'
+
+export interface FastifyListenOptions {
+  /** Default to `0` (picks the first available open port). */
+  port?: number
+  /** Default to `localhost`. */
+  host?: string
+  /** Ignored when `port` is specified. */
+  path?: string
+  /** Default to `511`. */
+  backlog?: number
+  exclusive?: boolean
+  /** Makes an IPC pipe readable for all users. */
+  readableAll?: boolean
+  /** Makes an IPC pipe writable for all users. */
+  writableAll?: boolean
+  /** Disable dual-stack behavior for an IPv6 TCP server. */
+  ipv6Only?: boolean
+  /** Close the listening server when this signal is aborted. */
+  signal?: AbortSignal
+  /** Resolves the text logged after the server starts. */
+  listenTextResolver?: (address: string) => string
+}
 
 type ErrorHandlerResult = any | Promise<any>
